backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
mongoengine==0.28.0
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0
email-validator==2.1.0
python-dateutil==2.8.2
